moving logs into types
