+++ packages/augur-sdk/src/index.ts
@@ -1,4 +1,4 @@
 export { Augur, UserSpecificEvent } from "./Augur";
-export *  from "./ethereum/types";
+export *  from "@augurproject/types/types/ethereum/logs";
 export { Provider } from './ethereum/Provider';
 export * from "./utils";
