+++ packages/augur-sdk/src/state/db/UserSyncableDB.ts
@@ -1,6 +1,6 @@
 import {SyncableDB} from "./SyncableDB";
 import {Augur} from "../../Augur";
-import {ParsedLog} from "../../ethereum/types";
+import {ParsedLog} from "@augurproject/types/types/ethereum/logs";
 import {DB} from "./DB";
 
 /**
