+++ packages/augur-types/types/ethereum/logs.ts
