+++ packages/ethersjs-provider/tsconfig.json
@@ -9,6 +9,9 @@
   ],
   "references": [
     {
+      "path": "../augur-types"
+    },
+    {
       "path": "../augur-artifacts"
     },
     {
